requirements.txt
fastapi
mysql-connector-python
uvicorn
bcrypt
pyjwt
python-dotenv
requests
python-dateutil
pydantic[email]